Update Version from int to long.

+++ Source/Example/Scenarios/S05_Read_from_stream.cs
@@ -23,8 +23,8 @@ namespace Example.Scenarios
                 .ToArray();
 
             var existent = await Stream.TryOpenAsync(Partition, default);
-	        var stream = existent.Found ? existent.Stream : new Stream(Partition);
-	        await Stream.WriteAsync(stream, default, events);
+            var stream = existent.Found ? existent.Stream : new Stream(Partition);
+            await Stream.WriteAsync(stream, default, events);
         }
 
         async Task ReadSlice()
@@ -44,7 +44,7 @@ namespace Example.Scenarios
             Console.WriteLine("If slice size is > than WATS limit, continuation token will be managed automatically");
 
             StreamSlice<EventEntity> slice;
-            var nextSliceStart = 1;
+            var nextSliceStart = 1L;
 
             do
             {
@@ -53,8 +53,8 @@ namespace Example.Scenarios
                 foreach (var @event in slice.Events)
                     Console.WriteLine("{0}:{1} {2}-{3}", @event.Id, @event.Version, @event.Type, @event.Data);
 
-                nextSliceStart = slice.HasEvents 
-                    ? slice.Events.Last().Version + 1 
+                nextSliceStart = slice.HasEvents
+                    ? slice.Events.Last().Version + 1
                     : -1;
             }
             while (!slice.IsEndOfStream);
@@ -74,10 +74,10 @@ namespace Example.Scenarios
 
         class EventEntity
         {
-            public int Id      { get; set; }
+            public int Id { get; set; }
             public string Type { get; set; }
             public string Data { get; set; }
-            public int Version { get; set; }
+            public long Version { get; set; }
         }
     }
 }
