+++ Source/Streamstone.Tests/Storage.cs
@@ -63,7 +63,7 @@ namespace Streamstone
             return connectionString ?? "UseDevelopmentStorage=true";
         }
 
-        public static StreamEntity InsertStreamEntity(this Partition partition, int version = 0)
+        public static StreamEntity InsertStreamEntity(this Partition partition, long version = 0)
         {
             var entity = new StreamEntity
             {
@@ -76,7 +76,7 @@ namespace Streamstone
             return entity;
         }
 
-        public static StreamEntity UpdateStreamEntity(this Partition partition, int version = 0)
+        public static StreamEntity UpdateStreamEntity(this Partition partition, long version = 0)
         {
             var entity = RetrieveStreamEntity(partition);
             entity.Version = version;
@@ -93,7 +93,7 @@ namespace Streamstone
 
         public static void InsertEventEntities(this Partition partition, params string[] ids)
         {
-            for (int i = 0; i < ids.Length; i++)
+            for (long i = 0; i < ids.Length; i++)
             {
                 var e = new EventEntity
                 {
